requirements.txt
appdirs==1.4.0
dj-database-url==0.4.1
Django==1.10.4
django-datetime-widget==0.9.3
gunicorn==19.6.0
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
pytz==2016.10
six==1.10.0
whitenoise==3.2
